move common warnings code to toplevel
